fixed issue from design feedback
